Rename nav package as profileNav
