+++ packages/profile-nav/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@bufferapp/publish-nav",
+  "name": "@bufferapp/publish-profile-nav",
   "version": "2.0.0",
   "description": "Buffer Nav Package",
   "main": "index.js",
